backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-skills-system"
version = "1.0.0"
description = "Agent Skills 知识问答+任务处理系统"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Agent Skills Team"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",

    # LangChain & AI
    "langchain>=0.3.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.12.0",

    # Database & Cache
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",

    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.1",

    # Vector Database
    "chromadb>=0.4.0",
    "sentence-transformers>=2.3.0",

    # Utilities
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "networkx>=3.2.1",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.5.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

all = [
    "agent-skills-system[dev,test]",
]

[project.urls]
Homepage = "https://github.com/your-org/agent-skills-system"
Documentation = "https://github.com/your-org/agent-skills-system/docs"
Repository = "https://github.com/your-org/agent-skills-system.git"
Issues = "https://github.com/your-org/agent-skills-system/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "core*", "models*", "config*", "scripts*", "tests*"]

# Black Configuration
[tool.black]
line-length = 99
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | \.eggs
    | \.tox
    | build
    | dist
    | __pycache__
)/
'''

# Ruff Configuration
[tool.ruff]
line-length = 99
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ASYNC", # flake8-async
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TID", # tidy imports
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.isort]
known-first-party = ["api", "core", "models", "config", "scripts"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# MyPy Configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "openai.*",
    "redis.*",
    "networkx.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["api", "core", "models", "config"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
